requirements.txt
uiautomation
pywin32
pillow
pytesseract
mss
pydantic
requests
pyyaml
opencv-python
pyautogui
psutil
